requirements.txt
streamlit==1.39.0
sounddevice==0.5.1
numpy==2.1.2
deepgram-sdk==v3.7.2
pydub==0.25.1
python-dotenv==1.0.1
scipy==1.14.1
fastapi==0.115.2
uvicorn==0.31.1
python-multipart==0.0.12
Jinja2==3.1.4